requirements.txt
datasets>=2.18.0
openai>=1.0.0
Pillow>=10.0.0
tqdm>=4.60.0
